route JS와 func JS의 확실한 분리 + 문서 설정 초기화 버튼 추가
https://github.com/openNAMU/openNAMU/issues/2101
