+++ route_go/main.go
@@ -59,6 +59,8 @@ func main() {
 		route_data = route.Api_bbs_w_comment(call_arg[1:])
 	} else if call_arg[0] == "api_bbs_w_n" {
 		route_data = route.Api_bbs_w(call_arg[1:])
+	} else if call_arg[0] == "api_w_set_reset" {
+		route_data = route.Api_w_set_reset(call_arg[1:])
 	} else {
 		log.Fatal("404")
 	}
